Bill_Transformation/windsurf-01/requirements.txt
# Core Dependencies
pandas>=1.5.0
openpyxl>=3.0.0
python-docx>=0.8.11
Jinja2>=3.0.0
pdfkit>=1.0.0
weasyprint>=53.0
python-dateutil>=2.8.2

# Development Dependencies
pytest>=7.0.0
black>=22.0.0
flake8>=4.0.0

# For Excel Processing
xlrd>=2.0.0
xlwt>=1.3.0

# For PDF Merging
PyPDF2>=2.0.0

# For ZIP Packaging
pyzipper>=0.3.5

# For LaTeX Support (if needed)
# pylatex>=1.4.1
# latexmk>=4.75
